+++ slack-market/commands/bought.rb
@@ -10,18 +10,18 @@ module SlackMarket
         quotes = Market.quotes(stocks) if stocks
         if quotes.any?
           quotes.each do |quote|
-            logger.info "#{client.owner}, user=#{user} - BOUGHT #{quote.name} (#{quote.symbol}): $#{quote.last_trade_price}"
+            logger.info "#{client.owner}, user=#{user} - BOUGHT #{quote.company_name} (#{quote.symbol}): $#{quote.latest_price}"
             if user.positions.where(symbol: quote.symbol, sold_at: nil).any?
-              client.say channel: data.channel, text: "#{user.slack_mention} already holds #{quote.name} (#{quote.symbol})"
+              client.say channel: data.channel, text: "#{user.slack_mention} already holds #{quote.company_name} (#{quote.symbol})"
             else
               Position.create!(
                 user: user,
                 purchased_at: Time.now.utc,
-                purchased_price_cents: quote.last_trade_price.to_f * 100,
+                purchased_price_cents: quote.latest_price.to_f * 100,
                 symbol: quote.symbol,
-                name: quote.name
+                name: quote.company_name
               )
-              client.say channel: data.channel, text: "#{user.slack_mention} bought #{quote.name} (#{quote.symbol}) at ~$#{quote.last_trade_price}"
+              client.say channel: data.channel, text: "#{user.slack_mention} bought #{quote.company_name} (#{quote.symbol}) at ~$#{quote.latest_price}"
             end
           end
         end