+++ slack-market/commands/sold.rb
@@ -10,14 +10,14 @@ module SlackMarket
         quotes = Market.quotes(stocks) if stocks
         if quotes.any?
           quotes.each do |quote|
-            logger.info "#{client.owner}, user=#{user} - SOLD #{quote.name} (#{quote.symbol}): $#{quote.last_trade_price}"
+            logger.info "#{client.owner}, user=#{user} - SOLD #{quote.company_name} (#{quote.symbol}): $#{quote.latest_price}"
             position = user.positions.where(symbol: quote.symbol, sold_at: nil).first
             if position
-              display = position.display(quote.last_trade_price.to_f * 100)
-              position.update_attributes!(sold_at: Time.now.utc, sold_price_cents: quote.last_trade_price.to_f * 100)
-              client.say channel: data.channel, text: "#{user.slack_mention} sold #{quote.name} at ~$#{quote.last_trade_price}, #{display}"
+              display = position.display(quote.latest_price.to_f * 100)
+              position.update_attributes!(sold_at: Time.now.utc, sold_price_cents: quote.latest_price.to_f * 100)
+              client.say channel: data.channel, text: "#{user.slack_mention} sold #{quote.company_name} at ~$#{quote.latest_price}, #{display}"
             else
-              client.say channel: data.channel, text: "#{user.slack_mention} does not hold #{quote.name} (#{quote.symbol})"
+              client.say channel: data.channel, text: "#{user.slack_mention} does not hold #{quote.company_name} (#{quote.symbol})"
             end
           end
         end