+++ spec/slack-market/commands/bought_spec.rb
@@ -18,20 +18,20 @@ describe SlackMarket::Commands::Bought do
       it 'records a buy', vcr: { cassette_name: 'msft' } do
         expect do
           expect(message: "#{SlackRubyBot.config.user} bought MSFT", user: user.user_id).to respond_with_slack_message(
-            "#{user.slack_mention} bought Microsoft Corporation (MSFT) at ~$84.26"
+            "#{user.slack_mention} bought Microsoft Corporation (MSFT) at ~$89.24"
           )
         end.to change(Position, :count).by(1)
         position = Position.last
         expect(position.symbol).to eq 'MSFT'
         expect(position.name).to eq 'Microsoft Corporation'
-        expect(position.purchased_price_cents).to eq 8426
+        expect(position.purchased_price_cents).to eq 8924
         expect(position.purchased_at).to_not be nil
       end
       context 'with an owned position' do
         let!(:position) { Fabricate(:position, user: user, name: 'XYZ Corporation', symbol: 'XYZ') }
         it 'does not support multiple buys' do
           allow(Market).to receive(:quotes).with([position.symbol]).and_return([
-                                                                                 OpenStruct.new(symbol: position.symbol, name: position.name, last_trade_price: 123)
+                                                                                 OpenStruct.new(symbol: position.symbol, company_name: position.name, latest_price: 123)
                                                                                ])
           expect do
             expect(message: "#{SlackRubyBot.config.user} bought #{position.symbol}", user: user.user_id).to respond_with_slack_message(